Made a lot of small changes to allow PHP to run with error_reporting E_ALL without giving notices. Most functions have been checked but some situations might give a notice.

+++ delete_domain.php
@@ -27,10 +27,10 @@ if (!level(5))
         
 }
 
-if ($_GET["id"]) {
-        if ($_GET["confirm"] == '0') {
+if (isset($_GET["id"])) {
+        if ((isset($_GET["confirm"])) && ($_GET['confirm'] == '0')) {
                 clean_page("index.php");
-        } elseif ($_GET["confirm"] == '1') {
+        } elseif ((isset($_GET["confirm"])) && ($_GET['confirm'] == '1')) {
                 delete_domain($_GET["id"]);
                 clean_page("index.php");
         }
